+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 import localFont from 'next/font/local';
 import './globals.css';
-import Header from '@/components/common/Header';
+import Header from '@/components/ui/Header';
 
 const pretendard_var = localFont({
   src: '../../public/fonts/pretendard_variable.woff2',
