+++ src/components/ui/NavItem.tsx
